Added UserController

+++ src/main/java/ua/skorobahatyi/onlineshopping/controller/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUser(@PathVariable Long id) {
+    public ResponseEntity<User> getUser(@PathVariable Integer id) {
         //logger.debug("GET /users/{}", id);
         User user = userService.getUser(id);
